requirements.txt
opencv-python
matplotlib
torch
pyyaml